pyproject.toml
[project]
name = "pyprom-exporters"
version = "0.0.1"
description = "."
readme = "README.md"
authors = [{ name = "Andreas A. Grammenos", email = "ag926@cl.cam.ac.uk" }]
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = ["prometheus-client>=0.20.0", "omegaconf>=2.3.0"]

[dependency-groups]
dev = [
    "ruff>=0.11.11",
    "pytest>=8.4.0",
    "pre-commit>=4.2.0",
    "pylint>=3.3.3",
    "python-kasa>=0.10.2",
]
tapo = ["python-kasa>=0.10.2"]

[project.scripts]
# kasa-probe = "pyprom_exporters.kasa_probe:main"
prom-exporter = "pyprom_exporters.prom_exporter:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 119
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Enable the isort rules.
select = ["F", "I"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []


# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas
# Note: if you enable this you should ignore COM812 formatter rule.
# skip-magic-trailing-comma = true

# Like Black, automatically detect the appropriate line ending.
# line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 119
# disable = """
#     too-many-arguments,
#     not-callable
# """

[tool.mypy]
check_untyped_defs = true
